+++ app/components/job/graphs/tasks-running-time-graph/tasks-running-time-graph.html
@@ -1,5 +1,2 @@
 <canvas blChart [datasets]="datasets" [options]="options" [chartType]="type" [colors]="colors">
 </canvas>
-<div class="loading-overlay" *ngIf="loading">
-    Loading tasks. This can take a long time.
-</div>
