+++ app/components/job/graphs/tasks-running-time-graph/tasks-running-time-graph.scss
@@ -4,22 +4,4 @@ bl-tasks-running-time-graph {
     display: block;
     width: 100%;
     height: 100%;
-
-    .loading-overlay {
-        $loading-overlay-width: 400px;
-        $loading-overlay-height: 70px;
-        position: absolute;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        top: calc(50% - #{$loading-overlay-height/2});
-        left: calc(50% - #{$loading-overlay-width/2});
-        margin: auto;
-        font-size: 16px;
-        width: $loading-overlay-width;
-        height: $loading-overlay-height;
-        background: $mineshaft-grey;
-        color: white;
-        opacity: 0.5;
-    }
 }
